app_dev/requirements.txt
requests>=2,<3
msal>=0.6.1,<2
python-dotenv
azure-search-documents==11.4.0
openai==1.8.0
tiktoken
python-docx
PyPDF2
tenacity
promptflow==1.5.0
promptflow-tools==1.2.0
promptflow-vectordb==0.2.4
keyrings.alt==5.0.0